Fix largefile issues in sample.c
Merge patch to fix issues when operating on large
files in sample_max_min() in sample.c.

Patch by Avuton Olrich


git-svn-id: svn://svn.code.sf.net/p/wavbreaker/code/trunk@284 1b2899f7-b92e-0410-a9e2-f6ce2bc8120a

+++ src/sample.c
@@ -344,12 +344,12 @@ void sample_close_file()
 static void sample_max_min(GraphData *graphData, double *pct)
 {
     int tmp = 0;
-    int ret = 0;
+    long int ret = 0;
     int min, max;
     int min_sample, max_sample;
-    int i, k;
+    long int i, k;
     int numSampleBlocks;
-    double tmp_sample_calc;
+    long int tmp_sample_calc;
     unsigned char devbuf[BLOCK_SIZE];
     Points *graph_data;
